Create an abstract class for metadata of multiple iceberg catalogs
