--- presto-iceberg/src/test/java/com/facebook/presto/iceberg/TestIcebergDistributedNative.java
@@ -1,123 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.iceberg;
-
-import com.facebook.presto.testing.MaterializedResult;
-import com.facebook.presto.testing.QueryRunner;
-import com.facebook.presto.tests.AbstractTestDistributedQueries;
-import com.google.common.collect.ImmutableMap;
-import org.testng.annotations.Test;
-
-import static com.facebook.presto.common.type.VarcharType.VARCHAR;
-import static com.facebook.presto.iceberg.CatalogType.HADOOP;
-import static com.facebook.presto.testing.MaterializedResult.resultBuilder;
-import static com.facebook.presto.testing.assertions.Assert.assertEquals;
-
-@Test
-public class TestIcebergDistributedNative
-        extends AbstractTestDistributedQueries
-{
-    @Override
-    protected QueryRunner createQueryRunner()
-            throws Exception
-    {
-        return IcebergQueryRunner.createNativeIcebergQueryRunner(ImmutableMap.of(), HADOOP);
-    }
-
-    @Override
-    protected boolean supportsViews()
-    {
-        return false;
-    }
-
-    @Override
-    protected boolean supportsNotNullColumns()
-    {
-        return false;
-    }
-
-    public void testJoinWithLessThanOnDatesInJoinClause()
-    {
-    }
-
-    @Override
-    public void testRenameTable()
-    {
-    }
-
-    @Override
-    public void testAddColumn()
-    {
-    }
-
-    @Override
-    public void testRenameColumn()
-    {
-    }
-
-    @Override
-    public void testDropColumn()
-    {
-    }
-
-    @Override
-    public void testInsert()
-    {
-    }
-
-    @Override
-    public void testCreateTable()
-    {
-    }
-
-    @Override
-    public void testDelete()
-    {
-    }
-
-    @Override
-    public void testShowColumns()
-    {
-        MaterializedResult actual = computeActual("SHOW COLUMNS FROM orders");
-
-        MaterializedResult expectedParametrizedVarchar = resultBuilder(getSession(), VARCHAR, VARCHAR, VARCHAR, VARCHAR)
-                .row("orderkey", "bigint", "", "")
-                .row("custkey", "bigint", "", "")
-                .row("orderstatus", "varchar", "", "")
-                .row("totalprice", "double", "", "")
-                .row("orderdate", "date", "", "")
-                .row("orderpriority", "varchar", "", "")
-                .row("clerk", "varchar", "", "")
-                .row("shippriority", "integer", "", "")
-                .row("comment", "varchar", "", "")
-                .build();
-
-        assertEquals(actual, expectedParametrizedVarchar);
-    }
-
-    @Override
-    public void testDescribeOutput()
-    {
-    }
-
-    @Override
-    public void testDescribeOutputNamedAndUnnamed()
-    {
-    }
-
-    @Override
-    public void testWrittenStats()
-    {
-    }
-}
